refactor: server socket to use direct connections
Updated ServerSocket to listen and handle a connection. It will keep
reading from that connection until the client disconnects. If there is
no data to read the server will wait 100ms and check again.

The update was done accordingly with how screen is sending data which is
done trough a single client for the whole connection.